+++ webknossos/webknossos/client/_generated/api/default/dataset_update.py
@@ -1,10 +1,10 @@
-from typing import Any, Dict
+from typing import Any, Dict, Union
 
 import httpx
 
 from ...client import Client
 from ...models.dataset_update_json_body import DatasetUpdateJsonBody
-from ...types import Response
+from ...types import UNSET, Response, Unset
 
 
 def _get_kwargs(
@@ -13,6 +13,7 @@ def _get_kwargs(
     *,
     client: Client,
     json_body: DatasetUpdateJsonBody,
+    skip_resolutions: Union[Unset, None, bool] = UNSET,
 ) -> Dict[str, Any]:
     url = "{}/api/datasets/{organizationName}/{dataSetName}".format(
         client.base_url, organizationName=organization_name, dataSetName=data_set_name
@@ -20,6 +21,11 @@ def _get_kwargs(
 
     headers: Dict[str, Any] = client.get_headers()
     cookies: Dict[str, Any] = client.get_cookies()
+
+    params: Dict[str, Any] = {
+        "skipResolutions": skip_resolutions,
+    }
+    params = {k: v for k, v in params.items() if v is not UNSET and v is not None}
 
     json_json_body = json_body.to_dict()
 
@@ -29,6 +35,7 @@ def _get_kwargs(
         "cookies": cookies,
         "timeout": client.get_timeout(),
         "json": json_json_body,
+        "params": params,
     }
 
 
@@ -47,12 +54,14 @@ def sync_detailed(
     *,
     client: Client,
     json_body: DatasetUpdateJsonBody,
+    skip_resolutions: Union[Unset, None, bool] = UNSET,
 ) -> Response[Any]:
     kwargs = _get_kwargs(
         organization_name=organization_name,
         data_set_name=data_set_name,
         client=client,
         json_body=json_body,
+        skip_resolutions=skip_resolutions,
     )
 
     response = httpx.patch(
@@ -68,12 +77,14 @@ async def asyncio_detailed(
     *,
     client: Client,
     json_body: DatasetUpdateJsonBody,
+    skip_resolutions: Union[Unset, None, bool] = UNSET,
 ) -> Response[Any]:
     kwargs = _get_kwargs(
         organization_name=organization_name,
         data_set_name=data_set_name,
         client=client,
         json_body=json_body,
+        skip_resolutions=skip_resolutions,
     )
 
     async with httpx.AsyncClient() as _client: